KOGITO-8604 - Saga examples use wrong groupId
