+++ kogito-quarkus-examples/process-saga-quarkus/pom.xml
@@ -6,7 +6,7 @@
     <artifactId>kogito-quarkus-examples</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <groupId>org.kie.kogito</groupId>
+
   <artifactId>process-saga-quarkus</artifactId>
 
   <name>Kogito Example :: Process Saga with Quarkus</name>
