--- kogito-quarkus-examples/process-saga-quarkus/src/main/java/org/kie/kogito/OrderService.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.kogito;
-
-import javax.enterprise.context.ApplicationScoped;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@ApplicationScoped
-public class OrderService {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(OrderService.class);
-
-    public Response success(String orderId) {
-        LOGGER.info("Order Success for order {}", orderId);
-        return Response.success(orderId);
-    }
-
-    public Response failure(String orderId) {
-        LOGGER.info("Order Failed for order {}", orderId);
-        return Response.error(orderId);
-    }
-}
