+++ kogito-quarkus-examples/process-saga-quarkus/src/main/java/org/kie/kogito/examples/PaymentService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito;
+package org.kie.kogito.examples;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
@@ -22,25 +22,25 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @ApplicationScoped
-public class ShippingService {
+public class PaymentService {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(ShippingService.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(PaymentService.class);
 
     @Inject
     MockService mockService;
 
-    public Response scheduleShipping(String orderId, String failService) {
-        LOGGER.info("Schedule Shipping for order {}", orderId);
-        return mockService.execute(failService, ShippingService.class, false);
+    public Response processPayment(String orderId, String failService) {
+        LOGGER.info("Process Payment for order {}", orderId);
+        return mockService.execute(failService, PaymentService.class, false);
     }
 
-    public Response scheduleShipping(String orderId, String failService, String throwException) {
-        LOGGER.info("Schedule Shipping for order {}", orderId);
-        return mockService.execute(failService, ShippingService.class, Boolean.parseBoolean(throwException));
+    public Response processPayment(String orderId, String failService, String throwException) {
+        LOGGER.info("Process Payment for order {}", orderId);
+        return mockService.execute(failService, PaymentService.class, Boolean.parseBoolean(throwException));
     }
 
-    public Response cancelShipping(String id) {
-        LOGGER.info("Cancel Shipping for  order {}", id);
+    public Response cancelPayment(String id) {
+        LOGGER.info("Cancel Payment for payment {}", id);
         return new Response(Response.Type.SUCCESS, id);
     }
 }
