+++ kogito-quarkus-examples/process-saga-quarkus/src/main/java/org/kie/kogito/examples/ServiceException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito;
+package org.kie.kogito.examples;
 
 public class ServiceException extends RuntimeException {
