+++ kogito-quarkus-examples/process-saga-quarkus/src/test/java/org/kie/kogito/examples/ProcessSagaIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito;
+package org.kie.kogito.examples;
 
 import org.junit.jupiter.api.Test;
