+++ serverless-workflow-examples/serverless-workflow-saga-quarkus/pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>org.kie.kogito</groupId>
+  <groupId>org.kie.kogito.examples</groupId>
   <version>1.0-SNAPSHOT</version>
   <artifactId>serverless-workflow-saga-quarkus</artifactId>
