+++ serverless-workflow-examples/serverless-workflow-saga-quarkus/src/main/java/org/kie/kogito/examples/PaymentService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito;
+package org.kie.kogito.examples;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
@@ -22,25 +22,25 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @ApplicationScoped
-public class StockService {
+public class PaymentService {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(StockService.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(PaymentService.class);
 
     @Inject
     MockService mockService;
 
-    public Response reserveStock(String orderId, String failService) {
-        LOGGER.info("Reserve Stock for order {}", orderId);
-        return mockService.execute(failService, StockService.class, true, orderId);
+    public Response processPayment(String orderId, String failService) {
+        LOGGER.info("Process Payment for order {}", orderId);
+        return mockService.execute(failService, PaymentService.class, true, orderId);
     }
 
-    public Response reserveStock(String orderId, String failService, String throwException) {
-        LOGGER.info("Reserve Stock for order {}", orderId);
-        return mockService.execute(failService, StockService.class, Boolean.parseBoolean(throwException), orderId);
+    public Response processPayment(String orderId, String failService, String throwException) {
+        LOGGER.info("Process Payment for order {}", orderId);
+        return mockService.execute(failService, PaymentService.class, Boolean.parseBoolean(throwException), orderId);
     }
 
-    public Response cancelStock(String id) {
-        LOGGER.info("Cancel Stock {}", id);
+    public Response cancelPayment(String id) {
+        LOGGER.info("Cancel Payment {}", id);
         return new Response(Response.Type.SUCCESS, id);
     }
 }
