+++ serverless-workflow-examples/serverless-workflow-saga-quarkus/src/main/java/org/kie/kogito/examples/Response.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito;
+package org.kie.kogito.examples;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
