+++ serverless-workflow-examples/serverless-workflow-saga-quarkus/src/main/java/org/kie/kogito/examples/ShippingService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito;
+package org.kie.kogito.examples;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
@@ -22,25 +22,25 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @ApplicationScoped
-public class PaymentService {
+public class ShippingService {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(PaymentService.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(ShippingService.class);
 
     @Inject
     MockService mockService;
 
-    public Response processPayment(String orderId, String failService) {
-        LOGGER.info("Process Payment for order {}", orderId);
-        return mockService.execute(failService, PaymentService.class, true, orderId);
+    public Response scheduleShipping(String orderId, String failService) {
+        LOGGER.info("Schedule Shipping for order {}", orderId);
+        return mockService.execute(failService, ShippingService.class, true, orderId);
     }
 
-    public Response processPayment(String orderId, String failService, String throwException) {
-        LOGGER.info("Process Payment for order {}", orderId);
-        return mockService.execute(failService, PaymentService.class, Boolean.parseBoolean(throwException), orderId);
+    public Response scheduleShipping(String orderId, String failService, String throwException) {
+        LOGGER.info("Schedule Shipping for order {}", orderId);
+        return mockService.execute(failService, ShippingService.class, Boolean.parseBoolean(throwException), orderId);
     }
 
-    public Response cancelPayment(String id) {
-        LOGGER.info("Cancel Payment {}", id);
+    public Response cancelShipping(String id) {
+        LOGGER.info("Cancel Shipping {}", id);
         return new Response(Response.Type.SUCCESS, id);
     }
 }
